lopper: add pretty printing for DTS output
re-work the node_dump routine (now called node_print), to better pretty
print the output of a FDT.

This is a stress test for the decode routines, as well as a way to put
symbolic phandles back into the output. The --pretty command line option
now controls whether dtc or the internal routine is used for writing
dts files.

This is a step along the way to CDO printing the topology.

Signed-off-by: Bruce Ashfield <bruce.ashfield@xilinx.com>