lopper: allow comments to be maintained through the pruning process
It has been noted that due to libfdt processing dtbs only, that
comments are lost through the lopping process.

We can mitigate this information loss by preprocessing the input
file and changing comments to node properties. When lopper pretty
prints a file, it knows how to translate those properties back
to comments.

Obviously the comments may not be accurate after translation ..
but that is an acceptable risk.

The preprocessing behaviour is only triggered via the --pretty
flag

Signed-off-by: Bruce Ashfield <bruce.ashfield@xilinx.com>