lopper: add 'noexec' and parent refcounting
Adding a new property to lops "noexec", which when present
means that the operation should be ignored. This is useful
for keeping lops in a tree, but not always having them used
(debug, etc).

We also extend the reference counting to opionally add a
reference to parent nodes when a child node is referenced.
This is useful when tracking tree wide node use.

A new method nodes_refd() is also added to easily get a list
of referenced nodes.

Signed-off-by: Bruce Ashfield <bruce.ashfield@xilinx.com>