openamp: add support for cpumask field
When processing cpu references, we should also consult the cpumask element
of the access property.

In this change, we factor the cpu processing to a separate routine, to make
the main openamp routine more readable.

In this new routine, we check the cpumask when processing cpu clusters. Not
only will we now delete un-referenced clusters, we'll delete usused cpu
nodes within a cluster (based on the mask).

Signed-off-by: Bruce Ashfield <bruce.ashfield@xilinx.com>