lopper: make phandle remapping more robust
Rather than using the phandle map directly, during pretty printing
we can walk the nodes and look for properties that are known to have
phandle references.

If we know the element to check (we do), we can lookup the target
node, extract its name and use that instead of the numerica value.

This allows us to adapt to renamed nodes and be more precise than
a regex in the replacement.

When printing nodes that have a non-zero phandle, we put out a label
so that any phandle references can find them if dtc is used in
future phases.

Signed-off-by: Bruce Ashfield <bruce.ashfield@xilinx.com>