lopper: drop invalid phandle references from output
When using "pretty" printed output, we do phandle replacement.
This means that we can also detect when a phandle is invalid (not
in the tree).

When we detect this, we can drop the elements that reference it,
since they are no longer valid.

When running in verbose mode, any dropped phandles will be
output.

Signed-off-by: Bruce Ashfield <bruce.ashfield@xilinx.com>