lopper: provide phandle utility routines
When processing a node, it is often the case that we want to
determine any phandles, and follow them to the referenced node.

This can be for validation, reference counting, or any number
of other uses.

The first use of this is the ability to find all references from
a given node. This is used by openamp to refcount the tree and
then remove any unref'd nodes.

Signed-off-by: Bruce Ashfield <bruce.ashfield@xilinx.com>