lopper: allow node regex to return paths
It is common that you want to match on a regex, and want the
paths and not the nodes. This is typically when you'll be deleting
nodes .. and node numbers can change, so you want paths.

We add a flag to the regex match routine so that it can return
paths and not just node numbers.

Signed-off-by: Bruce Ashfield <bruce.ashfield@xilinx.com>