lopper: split into lopper, FDT and tree components
In preparation for more complex lopper and library integration,
split lopper into multiple components. In particular we create
the following:

  - lopper.py: the existing lopper that takes command line options,
               manipulates a systemd device tree and generates output
  - lopper_fdt.py : the static methods for interfacing with a fdt
                    backed device tree
  - lopper_tree.py: Routines for representing, manipulating and
                    inspected a flattened device tree.

More splits will occur in the future (in particular FDT elements
will be moved out of lopper_tree.py, but this current split is the
amount of flexibility required for beta.

No changes have been made to the inputs (lops, device tree or
outputs), the changes are internal to lopper and loaded modules.

Signed-off-by: Bruce Ashfield <bruce.ashfield@xilinx.com>