assists/cdo: don't assume directory structure for import
The cdo assist is loading a cdotypes python file to define some
attributes. It had the "assists" subdirectory implied in that import
statement. Since modules can be anywhere, that isn't a valid
assumption.

We add the current file's path to the search directories, and let
the import loader locate it for us.

Signed-off-by: Bruce Ashfield <bruce.ashfield@xilinx.com>