lopper: restore binary SDT processing
Early versions of lopper could operate without the system
device tree being in source format.

Although less functionality is available with a binary system
device tree, it is still a possible use case.

With this commit, we restore the ability to take a binary SDT
and we tweak the cleanup routines to not delete it when finished
if that is the case.

Signed-off-by: Bruce Ashfield <bruce.ashfield@xilinx.com>