lopper: add property type guess
Add the ability to determine if a property is binary versus
always assuming that they are a list of numbers.

Since we don't (yet) have access to the dtb markers this guess
routine looks into the byte array and offers an opinion. The
logic here is similar to what dtc does when dumping a dtb.

Signed-off-by: Bruce Ashfield <bruce.ashfield@xilinx.com>