lopper: add 'conditional', 'code' lops
This unfortunately is a large commit, complete with re-factoring
and some embedded bug fixes. The changes were difficult to
make bisectable, so instead this is a complete: add new functionaliy
type commit.

No external interfaces have changed, but two new lopper operations
are avaible: conditional and code.

The details of the lops are in documented in the sanity and README
files.

The intent behind these operations is to give a light weight way
to manipulate/query the tree without needing to write a full blown
assists (since you then need to distribute the assist).

Signed-off-by: Bruce Ashfield <bruce.ashfield@xilinx.com>