lopper: add a meta data lopper operation
Add support for a meta data lopper operation, where the lop supplies
information to Lopper about how to translate/interpret the tree that
it is working on.

The first use of this is to allow a lop to update the phandle property
map, so that phandles can be symbolically replaced in an output dts
file.

Signed-off-by: Bruce Ashfield <bruce.ashfield@xilinx.com>