lopper/tree: allow variables to be passed into code block
Adding a 'env' dictionary to the exec_cmd call allows the
controlling flow to setup variables to be made available to the
block of code that will execute.

These are not intepreted by the exec_cmd, and are passed through
as-is and at the blocks own risk.

This allows us to create common blocks and pass context into
them for specific operations.

Signed-off-by: Bruce Ashfield <bruce.ashfield@xilinx.com>