lopper: add "select" lopper operation
To build up complex conditionals or series of nodes, we introduce
the "select" lopper operation.

The syntax of the test is exactly the same as the modify operation:

   <path to node>:<property>:<value>

If <path to node> is omitted, then the nodes from the previous
select operation are used. Hence you can build up a series of
tests to refine search operations.

Each select operation is a property of the format:

   select[_]<extension> = <select expression>

If "select"  with no extension is encountered, it clears any
previously selected nodes.

Operations are processed in the order they are found in the lop
node.

Once selected, other lopper operations will use the nodes if no
override is supplied in their lop. In this commit we convert
modify to do just that.

We also add __selected__ to the LopperTree class to hold a list
of selected nodes, and do some other minor cleanups to make the
variable names more explicit/clear.

Signed-off-by: Bruce Ashfield <bruce.ashfield@xilinx.com>