property: add subtype variable
We already know the composition of a properties values, since we
need it to property generate details during the resolve() function.

This can help us later compare and manipulate the values. So we
add a new member variable "subytpe" and we log the details there.
The LopperFmt types match this exactly, so we use the here.

We also convert the compare() method to use the subtype, rather
than superficially tested the elements during the test phase.

Signed-off-by: Bruce Ashfield <bruce.ashfield@xilinx.com>