lopper/tree: make property addition more robust
Enhance the ability to more simply add a property to a tree. When
you intially create a property, you don't always know the property
number (ok, you actually never do), the value or even the node
number it is assigned to.

So we provide defaults for all those values and we can fill them
in during resolution.

We also fixup the LopperNode's add() routine to assign itself to
a new property as the node. Without this, you can't ever resolve
the property.

Finally, we tweak LopperTreePrinter() to reset output to stdout
if a file hasn't been passed.

Signed-off-by: Bruce Ashfield <bruce.ashfield@xilinx.com>