lopper/tree: expand use of __must_sync__ flag
When a property is written to the backing FDT, numbers change. As
such, you need to do a full sync .. or operations as simple as looking
up a parent offset fail.

We set the __must_sync__ flag when we write a new property and expand
checking for it at critical treewide operations.

Signed-off-by: Bruce Ashfield <bruce.ashfield@xilinx.com>