lopper/sanity: fix selection2 test checking and add notes to node sync
We were looking for the wrong string in the selection 2 test, so it
was failing incorrectly.

Also, add some notes to the property add to existing node test. There
are all sorts of sync() calls you can make if you want, but they
are all optional now, since the tree will track and sort it out.

Signed-off-by: Bruce Ashfield <bruce.ashfield@xilinx.com>