lopper/tree: allow empty FDT trees to be created
To support adding a tree from scratch, we allow the FDT passed
to tree initialization to be None.

When detected, we'll initalize a new FDT and assign it to the
tree.

We also make the node add routine capable of handling the addition
of a Node AND any parent nodes that may not exist.

Signed-off-by: Bruce Ashfield <bruce.ashfield@xilinx.com>