lopper: convert output lop to LopperTree calls
Continue the abstraction of the libfdt calls and convert the lopper
output operation to use LopperTree calls.

The output is unchanged, but we have some bugfixes to deal with
issues around consistency when adding multiple nodes to a tree.

We also add the lnodes() call, to allow labels to be used as print
selectors.

Signed-off-by: Bruce Ashfield <bruce.ashfield@xilinx.com>