tree: track children as nodes, not paths
We were tracking child nodes as paths versus nodes. This was due
to the way the tree was originally constructed (we didn't have
the child nodes when resolving the parent). Now with the ability
to add nodes to the tree we have more flexibility and we can
track the nodes directly.

We also have some fallout and misc fixes due to the tracking
change.

Signed-off-by: Bruce Ashfield <bruce.ashfield@xilinx.com>