lopper: add yaml front end
Introduce yaml as a possible front (and backend) to lopper. This allows
specification of files in yaml format, which are converted into the
same internal representation as dts files, and hence can be any part
of the system (if properly formatted).

The dependency on anytree is optional, and if no yaml files are passed
lopper does not require it to be present.

Note: that this is currently for testing inputs and outputs and may
not work in all scenarios.

Signed-off-by: Bruce Ashfield <bruce.ashfield@xilinx.com>