lopper: throw exception when an assist can't be found
Python versions pre-3.6 don't raise an exception in resolve
when a file isn't found. So we can check for an empty resolved
path and throw the same exception. That way, we'll work in both
scenarios.

Also adding a checking of the Werror setting and exiting when
an assist isn't found.

Signed-off-by: Bruce Ashfield <bruce.ashfield@xilinx.com>