lopper: add "tree" lop
This lop works just like the output lop, except that the tree is
saved under a name, so that other lops can continue to modify
it.

We still have to change the other lops to work on it, but the
lop itself is functional.

Eventually most of the output lop can be removed and simply be
a follow on to this routine.

Signed-off-by: Bruce Ashfield <bruce.ashfield@xilinx.com>