lops/modify: allow node movement, not just rename
Allow the path of a node to be adjusted in a modify operation, not
just its name.

This allows us to move nodes around the tree, using the built in
tree primitives.

Also, since we can now target the tree, via the "tree" parameter
we can customize the location of node on a per-tree basis.

Signed-off-by: Bruce Ashfield <bruce.ashfield@xilinx.com>