assists: Add assists for generating cmake meta-data
Xilinx Baremetal Driver examples or few libraries(ex:lwip) and
it's examples supports a wide range of hardware features, All these
hardware features may not be present in a given system-dt and
application would like to know whether a particular feature is present
or not.

This patch adds support for the above use case by generating cmake
meta-data, this meta-data is consumed by the Xilinx Baremetal applications.

Signed-off-by: Appana Durga Kedareswara rao <appana.durga.rao@xilinx.com>