assists: baremetallinker_xlnx: Update memory handling based on the CPU cluster address-map property
In the System Devicetree processor memory range information is provided through
the address-map property in the cpu-cluster.

This patch updates the memory handling logic to use memory-range property
along with memory nodes, it loops over each memory node and searches for a
match in the address-map, if found corresponding data will be generated
in the linker else the memory node information will be ignored.

Signed-off-by: Appana Durga Kedareswara rao <appana.durga.rao@xilinx.com>