lopper: add rest API infrastructure
Adding the ability to launch a python flask REST server once lopper
has completed its run.

This is invoked when the --server flag is passed.

Note: This is NOT the final REST API, but is being added to show the
direction that lopper-as-a-server will go once the REST API is fully
defined and documented.

Signed-off-by: Bruce Ashfield <bruce.ashfield@xilinx.com>