lopper/yaml: ehance yaml typing and serialization
While the yaml -> json serialization is good for complex types
(like dictionaries) it is overly complex for simple types.

With this change, we look at the yaml node and if it is a string,
bool, int or list of simple types, we do not push it through
json, and it can be directly represented by the device tree.

Signed-off-by: Bruce Ashfield <bruce.ashfield@xilinx.com>