lopper: add 'xlate' lop and assist inheritance
It is becoming more common that non dts compatible trees / properties
are carried along side of device tree ones (i.e. yaml), and those
properties often need to be translated / expanded to be device tree
format (since they are complex types).

To make that easier, we add a new translate "xlate" lopper operation.
This lop is expected to work in coordination with a select lop to
target specific nodes and properties that need translation. xlate is
very similar to "code", and in fact, you can do everything that xlate
does with a code block (just more verbosely).

The differences (conveniences) from "code" are as follows:

  - Automatic inheritance of lopper library functions
  - Automatic iteration over selected nodes

The new property "inherit" is added to the code lop, and is hence
available to xlate. This propery is a comma separate list of python
assists that should be loaded and made available to the code block
on execution.

The python modules must be searchable by the python loader and in
an "assists" subdirectory. They will be loaded and made avaible under
their module name for direct use in the code block.

Examples follow:

                lop_0_1 {
                      compatible = "system-device-tree-v1,lop,select-v1";
                      select_1;
                      select_2 = "/domains/subsystem.*:cpus:.*";
                };
                lop_0_2 {
                        compatible = "system-device-tree-v1,lop,xlate-v1";
                        inherit = "subsystem";
                        code = "
                             subsystem.cpu_expand( tree, node )
                             subsystem.memory_expand( tree, node )
                             subsystem.access_expand( tree, node )
                        ";
                };

The impact of these two lops would be to run the code block in 0_2
against all nodes that match the selection criteria of 0_1. In this
case, it would be against any nodes under /domains/subsystem that have
a 'cpus' property (that is set to anything).

When the code block runs, the python module "subsystem" will be loaded
from the assists subdirectory and made available to the code block.
The calls to subsystem.<function> will leverage the transforms availble
in that assist (and in this case, will expand various properties in
in the node).

Signed-off-by: Bruce Ashfield <bruce.ashfield@xilinx.com>