lopper: allow permissive processing during property resolution
We don't always want to drop fields from properties if a phandle can't
be resolved (useful for debug, among other things). So we add a
"strict" flag to a tree, and only remove fields from properties if it
is set (and it is set by default).

We add the --permissive command line argument to lopper to switch on
the new option.

Signed-off-by: Bruce Ashfield <bruce.ashfield@xilinx.com>