lopper: trigger enhanced processing AFTER cpp is executed
Lopper can preserve comments and labels, but only in enhanced mode. We
are currently only doing the required preprocessing for enhanced mode on
the main system device tree file.

If we relocate the processing commands to the Lopper class, and trigger
them AFTER we execute cpp, then dtsi files and other source based
includes can also take advantage of this extra functionality.

Signed-off-by: Bruce Ashfield <bruce.ashfield@xilinx.com>