lopper: error on duplicate labels
We don't exit when dtc picks up errors due to the way that some partial
dts files are processed by lopper. As such, we weren't noticing when an
input file had duplicate labels, which are not valid and end up causing
issues with phandle replacement (among others).

We can easily pickup the duplicate labels when doing enhanced processing,
and dump information on resolving the issue.

Signed-off-by: Bruce Ashfield <bruce.ashfield@xilinx.com>