lopper: add the ability to select based on a property existing, or not
If a property is specified, without a value, then we are selecting
on whether or not that property exists (regardless of the value).

If we want to select nodes that do NOT contain a property, then the
name should be negated with a leading "!".

        lop_1 {
               compatible = "system-device-tree-v1,lop,select-v1";
               select_1;
               select_2 = "/.*:!phandle:";
        };

The above example would select all nodes that do not contain a
phandle property.

Signed-off-by: Bruce Ashfield <bruce.ashfield@xilinx.com>