treewide: fdt decoupling
Lopper started with a very simple goal: wrap libfdt to provide
facilities to modify a device tree.

It has now become a framework for many different things,
assists, analysis, transformation, format conversion, etc. To
provide even more advanced functionality we need to abstract
from libfdt in order to interface with different libraries.

This is a large, tree wide commit that completes the move of
all FDT processing to lopper_fdt. The visible/application LopperTree
functionality is unchanged (with compatibility functions
provided for some formerly required routines), but the internal
synchronization and construction routines no longer are coupled
to libfdt.

We now only exchange data with a FDT via a nested dictionary
structure, and the tree no longer needs to sync to the FDT
after changing nodes/properties. The tree is always consistent,
but can still be written to the FDT via a sync() and then
re-read into the tree.

This provides a 5x speed improvement, and simplifies many parts
of the code.

Signed-off-by: Bruce Ashfield <bruce.ashfield@xilinx.com>