assists: baremetalconfig: Pad Zero's in case of system device-tree node contains fewer address ranges
Baremetal pcie driver config struct contains multiple address properties
which usually specified via ranges property in the system device-tree,
For a given hardware configuration all the valid address ranges may not
be present in that case we need to pad zero's into the config struct to
avoid a compilation error.

This commit fixes this issue.

Signed-off-by: Appana Durga Kedareswara rao <appana.durga.rao@xilinx.com>