Add support for board topology cdo generation
This patch adds board topology cdo generation support to configure the
regulator rails.
We created a new lop file lop-versal-regulator-cdo.dts which contains
options argument as "regulator topology_board.cdo".
Here the options[0] which is regulator is to differentiate which cdo
the user is asking about and
the options[1] which is topology_board.cdo is the cdo file name.

The input to the lopper is system device tree which contains i2c
regulators information which will be parsed by cdo_topology.py.

The topology_headr.py contains all PM command and other macros required
to generated CDO data.

Signed-off-by: Naga Sureshkumar Relli <naga.sureshkumar.relli@xilinx.com>