lopper: add xlate command option
Add the ability to indicate a lop filed that should be executed for a
given category of input.

This can be specified multiple times and is additive:

  -x "*.yaml:my-yaml-xlate.dts" -x "*.dts:my-dts-xlate.dts"

The lop file specification (the part after the ":") is optional, and
if not passed one of the format:

  lop-xlate-<extension>.dts

Will be generated and added to the input files.

Once specified these are processed like any other input file and will
apply operations to the tree.

Note: these are not currently restricted to files of a given input
      type, but may be in the future (hency why they are advertised
      as such for now).

Signed-off-by: Bruce Ashfield <bruce.ashfield@xilinx.com>