lopper: phandle resolution / output fixues
It was observed that phandles which could not be resolved were not
property dropping fields from their related definitions.

This was due to incorrect processing in resolution. We re-work the
phandle processing in this routine, and manipulate the "records"
differnetly, so we can have consistent dropping of invalid phandles
and their associated fields.

Signed-off-by: Bruce Ashfield <bruce.ashfield@xilinx.com>