lopper: fix comment enhanced processing
Lopper can maintain comments through processing, by creating attributes
and allowing them to be compiled into the device tree.

As we introduced 'pcpp' support, more comments are flowing into the
consolidated device tree, and that has exposed that not all comments are
in valid locations to allow the device tree to be compiled.

We could (and will) relocate those comments to safe places and specific
comment nodes. But for now, if a comment is tagged as invalid, we'll
simply drop it, to ensure that processing can continue.

Signed-off-by: Bruce Ashfield <bruce.ashfield@xilinx.com>