lopper_tree: don't change property type if a string looks like a number
The property types that we get from from the export -> dictionary import
are better than they used to be (i.e. they are correct). So we shouldn't
be changing the entire property type when we can convert a string to an
integer.

We keep the type switch around, and commented out .. in case some hex
numbers pop up and cause us issues.

Signed-off-by: Bruce Ashfield <bruce.ashfield@xilinx.com>