lopper: pass --permissive to trees and property resolution
The permissive flag wasn't properly toggling the strict flag in
trees, and hence wasn't being used to dictate whether strict property
resolution was done.

Without this propagation, we have no way to allow a phandle node that
doesn't resolve to persist to the final output. There are cases where
this is something we need to do (debug, etc).

Signed-off-by: Bruce Ashfield <bruce.ashfield@xilinx.com>