tree: don't delete subnodes from parent
When re-homing a node that has subnodes, we were deleting all links
to the nodes from the starting node. That means we cannot snip a
subtree and move it to a new location (since the children links are
gone).

We can still delete the nodes from the tree, but maintain this linkage
so an entire subtree can be moved.

Signed-off-by: Bruce Ashfield <bruce.ashfield@xilinx.com>