lopper_tree: fix phandle replacement for multi record elements
We were previously comparing if the current target matched one of
the logged phandle resolutions. This will appear to mostl work,
until you get phandles and values that match.

for instance a phandle of 0x3 and a mask of 0x3. A phandle resolution
would match both of those and incorrectly do a symbolic replacement.

We should have always just compared if the current element in the
record was a phandle target, and done the replacement then.

Signed-off-by: Bruce Ashfield <bruce.ashfield@xilinx.com>