lopper_tree: don't change property encoding on __deepcopy__
When a node is deep copied, the properties are also copied at the
same time. We were directly assigning the copy via the value
attribute, which in turn has a convenience function to put values
into lists (now that we are no longer bound to libfdt, it is
debatable if want that behaviour at all).

That operation can change the encoding of some values (in particular
if they are json encoded strings). To avoid the function, we use
__dict__ and directly assign to the value.

Signed-off-by: Bruce Ashfield <bruce.ashfield@xilinx.com>