lopper: use python-device-tree as alternate parser
This commit furthers the split of lopper from libfdt. With this
change, we can now use the --no-libfdt option and use the dtlib
python device tree parser instead of dtc/libfdt.

This requires some factoring of the tree, but existing "Lopper"
interfaces are maintained through a global variable (versus
using the "Lopper" class).

We now have a lopper_base class that encapsulated manipulation
routines that are common to any front/backend. For broad compatibility
only these should be used by generic routines. Over time, more
common methods may be added.

Note: that when using a non-libdt parser, we can no longer sync
to a fdt, and hence cannot write dtbs. This functionality will
be added in the future.

Signed-off-by: Bruce Ashfield <bruce.ashfield@xilinx.com>