lopper: ensure system device tree dts file directory is in include paths
When we concatenate/join input dts (and yaml) files into the main device
tree for lopper, we create a file in /tmp/ to contain the results.

This happens before cpp by design.

As such, if the original dts file has #includes, etc, they won't be found
since they are likely in the path of the original SDT file, and not
in the tmp directory.

Before we call pcpp or cpp, we should grab the path of the SDK file and
ensure it can be searched.

Signed-off-by: Bruce Ashfield <bruce.ashfield@xilinx.com>