lopper_tree: abstract json properties
When a LopperProperty was created from yaml or json, the .value
class property is a json encoded string.

This string is not obvious, and must be json.loads() before it is
useful at all.

We abstract this detail, by providing index magic functions to the
LopperProperty.

When these routines are called, the origin of the property is
checked, and the json loaded on the fly and a structure returned.
If the origin was not json, then the standard value is returned.

The caller should simply check the type of the returned value and
operate accordingly.

Signed-off-by: Bruce Ashfield <bruce.ashfield@xilinx.com>