lopper_tree/subsystem: robustness changes for json processing
The subsystem *_expand() routines can be called with different inputs,
and are not always yaml.

We make them more robust by checking for the dictionary structure of
properties, and otherwise setting defaults or returning early.

We also tweak the default length of a LopperProperty to be 1, since
otherwise tests like:

   if prop:

Will no longer compare to None, but will use the lenght and the return
of 0 means a valid property will be skipped.

Signed-off-by: Bruce Ashfield <bruce.ashfield@xilinx.com>