lopper_yaml: support anchors and aliases
When reading a yaml file with anchors and aliases, information was being
lost since the alias would clobber the original value. This was due to the
fact that the dictionaries were references, so when a node was named, it
updated both copies.

The result: two nodes with the same name, and only one makes it out to
the Lopper tree.

We do a deepcopy of the dict, to make sure that the value really is
unique.

Also, we change the encoding of boolean values to be 1 or 0, so that they
will always appear in the LopperTree for processing.

Signed-off-by: Bruce Ashfield <bruce.ashfield@xilinx.com>