lopper: allow node merge to update properties
The first implementation of merge was not capable of updating properties
as part of a merge (only adding them). With this change, a node can be
allowed to update properties as part of a merge.

We also document the new node merge code, and document the lopper_base
class for clarity.

Signed-off-by: Bruce Ashfield <bruce.ashfield@xilinx.com>