lopper: move various files to lopper/ subdirectory
This is step 1 towards making lopper something we can package for
pypi.

Since lopper is split into multiple files, these must be in a
'package' in the 'python package with __init__.py' sense of the word.

Once that's done, the results can be 'packaged' in the sense of
'adding all the boilerplate files we need to create a wheel that can
be installed with pip' sense of the word.

This commit can't be run directly, because various imports need to be
fixed up next. That breaks bisectability, but I'm leaving it this way
for readability during review. I can squash the import fixes that
follow in the next commit into this patch if bisectable history is a
priority.

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>