lopper: extract __main__ from __init__
The norm in the python packaging world is that if a package is
runnable as a module, that's done with 'python -m <packagename>'. This
works under the hood by finding packagename/__main__.py and running
that.

Right now, that doesn't work for lopper because a previous patch moved
the lopper.py module into the main lopper/__init__.py file in the
newly created lopper package.

This patch restores the ability to run lopper from the command line by
extracting the relevant pieces into a new file, and adjusting the
imports accordingly. This is also a good opportunity to remove unused
imports that have piled up in __init__.py over time.

I'm hard-coding the name 'lopper' in usage() here.

This is also a prep work patch for creating a pypi package, which will
allow us to write an entry point script that calls code in
__main__.py.

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>