lopper: adjust __main__ to create an entry point
The split between main() and the code that follows it can be unified
into a single main() function without behavioral changes.

Doing that allows us to use the newly "unified" main() as an entry
point function to the package. This will be useful when we create a
pypi package, as it will allow us to make a 'lopper' executable that
calls main() and will be placed on PATH by 'pip install', giving a
cross-platform command-line entry point.

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>