lopper: prepare for single-sourcing the version
Create a VERSION file and use it to read the version. This will allow
us to read the same file from setup.py and using it as the version
when creating a wheel.

The current version format is not PEP 440 compliant, though. That
ought to be addressed before releasing a package by bumping the
version to something in PEP 440 style.

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>