lopper: make flask dependency optional to support integration
Reduce the mandatory library dependencies for lopper to enable it to run
natively in OpenEmbedded and Yocto build environments for device tree
processing during builds. This retains the ability to run the flask
server when the libraries are available.

Making flask support optional enables lopper to run using only the
python recipes that already support native builds in meta-virtualization
and its layer dependencies.

Signed-off-by: Christopher Clark <christopher.clark@starlab.io>