lopper: fix fdt/dtb write with empty tree
As part of the split of libfdt to dtlib, we now start dtb writes from
an empty FDT versus a copy of an existing one. This means that with some
larger properties three resize tries may not be enough. We increase this
to 10, and add a log message so we'll know if a property fails to write
to the FDT.

Signed-off-by: Bruce Ashfield <bruce.ashfield@xilinx.com>