lopper: allow loaded assists to import other assists by deafult
When we are running exec_cmd() assists can be imported, and those
assists may then include/import yet more assists.

secondary imports can fail, due to the 'assists' directory not
being on the system search path. If we add ./assists by default,
they will be found.

We also have a set of load paths that are passed to the lopper
front end, we can pass them into the exec_cmd() and use them as
even more load directories.

Signed-off-by: Bruce Ashfield <bruce.ashfield@xilinx.com>