lopper_tree: make assist loading path absolute
The previous attempt at fixing the issue where an assist that imports
another assists fails due to path issues .. was naive, in that it
assumed lopper was directly invoked from where the main script resides.

By using the path of __file__, we can avoid the working directory issues
and ensure that we can add our asissts subdirectory as absolute to the
python system path.

Signed-off-by: Bruce Ashfield <bruce.ashfield@xilinx.com>