lopper_fdt: reverse node order when syncing
When we sync nodes via libfdt, renumbering occurs. If we want a sibling
node to have a lower number and preserve ordering, we must reverse the
exported node dictionary (and properties) when syncing them to the FDT.

Newer nodes bump older ones to higher numbers when they share a parent.
So we work bottom up, to keep the order consistent.

Signed-off-by: Bruce Ashfield <bruce.ashfield@xilinx.com>