lopper_fdt: make node addition more robust
The attempt to fix node ordering on newly created FDTs exposed bugs in
the way that parent nodes were adding during sync .. in particular the
fact that they could only be one deep before breaking the routines.
(Three+ deep nodes were added to the root node).

With this change, we walk the path of any node addition and make sure
that all the parent components exist, before adding the node itself.

Signed-off-by: Bruce Ashfield <bruce.ashfield@xilinx.com>