lopper: add configuration file and value support
For lopper components that execute before lops are processed, there
needs to be an alternate way to specify configuration options. Core
Lopper can also leverage a configuration file, versus command line
options as appropriate (but for now, it is only used by the yaml
front end).

To support this, we introduce two new options:

   --cfgfile
   --cfgval

The configuration file and options are specified in Python
configparse format (.ini -style).

We also add a default lopper.ini in the same directory as the
main lopper executable to provide defaults.

This is an introductory commit, and currently only the yaml
section is supoorted, with a single option "bool_as_int".

Examples of the parameters follow:

  % lopper.py -f --configval=yaml.bool_as_int --enhanced -x '*.yaml' firewall-ref.yaml firewall-out.yaml

  % lopper.py -f --cfgfile /tmp/lopper.ini  --enhanced -x '*.yaml' firewall-ref.yaml firewall-out.yaml

Signed-off-by: Bruce Ashfield <bruce.ashfield@xilinx.com>