lopper_yaml: allow encoding of lists as nodes
If an imported dictionary has a list attribute, and all of the subtypes are dictionaries
we had yaml something like:

    firewallconf:
       - block: true
       - block: never
         domain: 0x1
       - ....

We can expand that, and generate node names for the subdictionaries. If we don't,
the list will be json encoded and will have to be expanded later.

The node expanded case results in:

                        firewallconf_0 {
                                block = <0x8>;
                        };
                        firewallconf_1 {
                                domain = "subsystem2";
                                block = "never";
                        };
                        firewallconf_2 {
                                domain = "subsystem3";
                                block = "never";
                        };

While the json encoded is:

                        firewallconf = "[{"block": 8}, {"domain": "subsystem2", "block": "never"}, {"domain": "subsystem3", "block": "never"}]";

This commit creates an optional toggle (set to json encoding for now),
so we can decide which is easier processing.

Signed-off-by: Bruce Ashfield <bruce.ashfield@xilinx.com>